+++ examples/my-app/CHANGELOG.md
@@ -1,5 +1,16 @@
 # my-app
 
+## 1.15.0
+
+### Minor Changes
+
+- Support sponsorship gas
+
+### Patch Changes
+
+- Updated dependencies
+  - @0xsodium/provider@1.16.0
+
 ## 1.14.0
 
 ### Minor Changes