+++ packages/transactions/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @0xsodium/transactions
 
+## 1.16.0
+
+### Minor Changes
+
+- Support sponsorship gas
+
+### Patch Changes
+
+- Updated dependencies
+  - @0xsodium/abi@1.16.0
+  - @0xsodium/network@1.16.0
+  - @0xsodium/utils@1.16.0
+
 ## 1.15.0
 
 ### Minor Changes